requirements.txt
tensorflow==2.18.0
keras==3.8.0
fastapi
uvicorn[standard]
pillow
numpy
requests
python-multipart
